requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
Pillow==7.2.0
python-slugify==4.0.1
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
